requirements.txt
mypy==1.7.0
python-dotenv==1.0.0